requirements.txt
ollama
srt
tqdm
PyQt6
opencc-python-reimplemented
